Move test_modularize_instance into test_core.py. NFC
Move this test alongside its peer (test_esm_integration).  These are
the only tests we have for these modes and its useful to be able to
test other modes such as wasm64.
